Add ROS Environment Widget
