+++ tesseract_rviz/CMakeLists.txt
@@ -338,13 +338,14 @@ find_package(tesseract_widgets REQUIRED)
 
 # Environment Panel
 qt_wrap_cpp(${PROJECT_NAME}_environment_plugin_core_cpp_MOCS
-  include/tesseract_rviz/environment_plugin/environment_panel.h
+  include/tesseract_rviz/environment_plugin/ros_environment_widget.h
+  include/tesseract_rviz/environment_plugin/rostopic_combo_box.h
   include/tesseract_rviz/environment_plugin/environment_display.h
-  include/tesseract_rviz/environment_plugin/rostopic_combo_box.h)
+)
 add_library(${PROJECT_NAME}_environment_plugin_core SHARED
-  src/environment_plugin/environment_panel.cpp
-  src/environment_plugin/environment_display.cpp
+  src/environment_plugin/ros_environment_widget.cpp
   src/environment_plugin/rostopic_combo_box.cpp
+  src/environment_plugin/environment_display.cpp
   src/environment_plugin/conversions.cpp
   ${${PROJECT_NAME}_environment_plugin_core_cpp_MOCS})
 target_link_libraries(${PROJECT_NAME}_environment_plugin_core PUBLIC
