+++ tesseract_rviz/include/tesseract_rviz/environment_plugin/environment_display.h
@@ -28,14 +28,6 @@ private Q_SLOTS:
   void onEnvironmentTopicChanged();
   void onJointStateTopicChanged();
 
-  void onEnvironmentSet(const tesseract_environment::Environment& env);
-  void onEnvironmentChanged(const tesseract_environment::Environment& env);
-  void onEnvironmentCurrentStateChanged(const tesseract_environment::Environment& env);
-  void onLinkVisibleChanged(const std::string& link_name, bool visible);
-  void onLinkCollisionVisibleChanged(const std::string& link_name, bool visible);
-  void onLinkVisualVisibleChanged(const std::string& link_name, bool visible);
-  void onSelectedLinksChanged(const std::vector<std::string>& selected_links);
-
   void onSceneAlphaChanged();
   void onSceneVisualVisibleChanged();
   void onSceneCollisionVisibleChanged();
@@ -44,14 +36,8 @@ private Q_SLOTS:
 protected:
   // overrides from Display
   void onInitialize() override;
-  void onEnable() override;
-  void onDisable() override;
-  void onUpdate();
 
   std::unique_ptr<EnvironmentDisplayPrivate> data_;
-
-  void clear();
-  void clearContainer(const tesseract_gui::EntityContainer& container);
 };
 }  // namespace tesseract_rviz
 #endif  // TESSERACT_RVIZ_ENVIRONMENT_DISPLAY_H
